Solucionado problema de obtener mas usuarios para la Geolocalizacion inicial. Falta el editar perfil, mensaje de informacion de venta de articulo favorito y los pm para terminar de forma inicial la aplicacion en el frontend.

+++ application/sql/bd.sql
@@ -29,8 +29,8 @@ create table usuarios(
                                                         on update cascade,
     activado            bool         not null default true,
     baneado             bool         not null default false,
-    latitud             bigint default null,
-    longitud            bigint default null
+    latitud             double precision default null,
+    longitud            double precision default null
 );
 
 drop table if exists tokens cascade;
